% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_importing.R
\name{combine_content}
\alias{combine_content}
\title{Scrape content from multiple URLs and combine into a single data frame}
\usage{
combine_content(n_max, ...)
}
\arguments{
\item{n_max}{Max number of rows to return. Please keep in mind your available
system resources and any potential burden on other servers when determing
the number of rows you need.}

\item{...}{Additional arguments to pass to construct_url to build api query.}
}
\value{
A data.frame with content imported from your query
}
\description{
Scrape content from multiple URLs and combine into a single data frame
}
