% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_building.R
\name{construct_url}
\alias{construct_url}
\title{Construct a URL to access Reddit Content}
\usage{
construct_url(type = "comment", ...)
}
\arguments{
\item{type}{Type of content you're looking for. Permissible values are "comment" and "submission"}

\item{...}{Additional parameters to pass to api query. See details below.}
}
\value{
A string containing a URL pointing to JSON Reddit data
}
\description{
Construct a URL to access Reddit Content
}
\details{
Here's the list of parameters that the api can use: https://github.com/pushshift/api#search-parameters-for-comments. The api ignores unrecognized parameters, so you may want to verify your output.
}
