% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_comments}
\alias{get_comments}
\title{Get comments from Reddit post}
\usage{
get_comments(elements)
}
\arguments{
\item{elements}{Object created by \link{\code{get_elements}}}
}
\value{
Tibble with one row for each comment, with all available attributes spread across
columns
}
\description{
Get comments from Reddit post
}

