% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_importing.R
\name{get_content}
\alias{get_content}
\title{Import data from URL}
\usage{
get_content(url)
}
\arguments{
\item{url}{A url leading to JSON data, typically produced from construct_url}
}
\value{
A data.frame with scraped data from the provided url
}
\description{
This function is a wrapper around jsonlite::fromJSON that returns
a data.frame instead of a list.
}
\examples{
# Read 25 recent Reddit comments into R
get_content(construct_url())
}
