% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_elements}
\alias{get_elements}
\title{Parse page into individual elements}
\usage{
get_elements(page)
}
\arguments{
\item{page}{Object created by \link{\code{get_page}}}
}
\value{
\code{\link[tidyjson]{tbl_json}} with one row for each element
of the page.
}
\description{
Parse page into individual elements
}

