% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_page_content}
\alias{get_page_content}
\title{Get Page Content}
\usage{
get_page_content(link)
}
\arguments{
\item{link}{Character string with the url of the Reddit page from which
to extract content.}
}
\value{
A character string containing the json data from the Reddit API call
}
\description{
Get Page Content
}
\examples{
get_page_content("http://www.reddit.com/")
get_page_content("http://www.reddit.com/r/politics/")
get_page_content("https://www.reddit.com/r/funny/comments/4xuzyl/")
get_page_content("https://www.reddit.com/r/funny/comments/4xuzyl/usain_bolt_celebrating_his_olympic_victory_with/")
}

