% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-importing.R
\name{import_reddit_content_from_url}
\alias{import_reddit_content_from_url}
\title{Import Data from a Pushshift URL}
\usage{
import_reddit_content_from_url(url)
}
\arguments{
\item{url}{A url leading to JSON data, typically produced via
\code{construct_pushshift_url}.}
}
\value{
A data.frame with data available from the provided url
}
\description{
This function is a wrapper around \code{jsonlite::fromJSON}
that returns a data.frame instead of a list. This function is primarily
useful for powering \code{get_reddit_content}, but can also be called
directly on a single url obtained via \code{construct_pushshift_url}.
}
\examples{
# read the 25 recent Reddit comments into R
import_reddit_content_from_url(construct_pushshift_url())

# construct a url using pushshift api parameters
comments_url <- construct_pushshift_url(
  content_type = "comment",
  q = "Stardew Valley",
  subreddit = "games"
)

# import data from that url
comments_df <- import_reddit_content_from_url(comments_url)
}
