% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/login_reddit.R
\name{login_reddit}
\alias{login_reddit}
\title{Login to reddit}
\usage{
login_reddit(user, pass, client, secret)
}
\arguments{
\item{user}{user name}

\item{pass}{password for user}

\item{client}{client ID}

\item{secret}{secret key}
}
\description{
Returns a token needed for API commands that require oauth
}

