% Generated by roxygen2 (4.0.2): do not edit by hand
\name{post_api}
\alias{post_api}
\title{Post Reddit API function}
\usage{
post_api(url, params, token)
}
\arguments{
\item{url}{URL of the API function}

\item{params}{A list of parameters to feed the API function}

\item{token}{oAuth token created by \code{\link{reddit_login()}} function}
}
\value{
Json object
}
\description{
Simple helper function to use any API function requring a POST method
}
\note{
See \url{http://www.reddit.com/dev/api/} for list of functions and required paramters for each API function
}
\examples{
# Post a comment to a thread
api_url <- "https://oauth.reddit.com/api/comment"
plist   <- list(text = "Hi Mom", thing_id = "t3_2is85o", api_type = "json")

post_api(url = api_url, params = plist, token = rtoken)
}

