% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_grabber.R
\name{post_grabber}
\alias{post_grabber}
\title{Grab Posts}
\usage{
post_grabber(url, nmult = 1, pnum = 100)
}
\arguments{
\item{url}{link to frontpage ("http://www.reddit.com/") or a subreddit ("http://www.reddit.com/r/nfl).}

\item{nmult}{Multipler applied to number of posts to grab (for 500 posts: nmult = 5, pnum = 100)}

\item{pnum}{Number of posts to grab in each set (25:100)}
}
\description{
Grabs posts and post information from a sub-reddit or the main page.
}
\note{
You can only grab up to the 1,000 most recent posts.
}

