% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_subreddit.R
\name{scrape_subreddit}
\alias{scrape_subreddit}
\title{Scrapes posts from a subreddit using pagination.}
\usage{
scrape_subreddit(subreddit, post_limit = 500)
}
\arguments{
\item{subreddit}{a string that contains the subbredit name,
the format 'r/name' or a complete Reddit URL. It is validated using \code{validate_subreddit}.}

\item{post_limit}{The maximum number of posts that are going to be scraped.
(Defaut value: 500).}
}
\value{
A \code{tibble} that contains the columns \code{title} y \code{selftext}
  with the data from the posts. The number of rows is lower or equal than \code{post_limit}.
}
\description{
This function calls the Reddit API in an iterative way (pagination) to surpass the 100-post limit 
per request and collects until the specified limit.
}
\examples{
\dontrun{
# Scrape the hottest first 250 posts from r/rstats
rstats_data <- scrape_subreddit("r/rstats", post_limit = 250)
head(rstats_data)
}
}
