% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{setup_reddit_oauth}
\alias{setup_reddit_oauth}
\title{Setup Reddit OAuth2.0}
\usage{
setup_reddit_oauth(key, secret)
}
\arguments{
\item{key}{Key generated from Reddit developed application}

\item{secret}{Secret generated from Reddit developed application}
}
\value{
Generates a \code{\link[httr]{Token}} to authenticate Reddit API calls
}
\description{
Setup Reddit OAuth2.0
}

