% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{split_type}
\alias{split_type}
\title{Split page into separate content types}
\usage{
split_type(page)
}
\arguments{
\item{page}{Object created by `get_page`}
}
\value{

}
\description{
Split page into separate content types
}

