% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{spread_attributes}
\alias{spread_attributes}
\title{Extract attributes from Reddit "things"}
\usage{
spread_attributes(elements)
}
\arguments{
\item{elements}{}
}
\value{
A \code{\link[tibble]{tibble}} with one row for each element,
with all available attributes spread across columns.
}
\description{
Extract attributes from Reddit "things"
}

