% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_words.R
\name{top_words}
\alias{top_words}
\title{Identifies the most frequent words in the Reddit posts.}
\usage{
top_words(reddit_tibble, n = 500)
}
\arguments{
\item{reddit_tibble}{A \code{tibble} that contains at least the columns
\code{title} y \code{selftext} (eg. the result from \code{scrape_subreddit}).}

\item{n}{The number of the most frequent words to return. (Default value: 500).}
}
\value{
A \code{tibble} with the columns \code{word} (word) and \code{n} (freequency),
  sorted from most to least frequency.
}
\description{
This function combines the titles and body of the posts, tokenizes the text,
removes 'stopwords' and short words, and calculates the N most used words.
}
\examples{
\dontrun{
# Assuming that 'data_scraped' is the result from scrape_subreddit
data_scraped <- tibble::tibble(
  title = c("R programming is great", "Python is better, test"), 
  selftext = c("I love R, I do", "Test, test, test.")
)
top_10_words <- top_words(data_scraped, n = 10)
print(top_10_words)
}
}
