% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_subreddit.R
\name{validate_subreddit}
\alias{validate_subreddit}
\title{Validates and extracts the name of a subreddit.}
\usage{
validate_subreddit(url_or_name)
}
\arguments{
\item{url_or_name}{A string that contains the name of the subreddit,
the format'r/name' or a complete Reddit URL.}
}
\value{
The clean name of the subreddit (just the name), if it is valid and if it exists.
  Stops the execcution with an error if the subdreddit does not exist or if it is invalid.
}
\description{
This function verifies if the input is a valid URL,'r/name' or just 'name'.
Also calls the reddit API to verify that the subreddit exists.
}
\examples{
\dontrun{
validate_subreddit("rprogramming")
validate_subreddit("https://old.reddit.com/r/dataisbeautiful/")
}
}
