% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-conversion.R
\name{api_to_date}
\alias{api_to_date}
\title{Convert API Time Back to a Timestamp}
\usage{
api_to_date(api_date, tz = "")
}
\arguments{
\item{api_date}{A date or vector of dates to be converted}

\item{tz}{The timezone you want the resulting output to be in, defaults
to system's timezone when calling this function directly.}
}
\value{
A POSIXct of the input
}
\description{
This is a convenience function that wraps \code{as.POSIXct} to
convert a time represented in seconds after the unit epoch to
a more human interpretable timestamp.
}
\examples{
# capture the current time
right_now <- Sys.time()

# convert it to api format (seconds after unit epoch)
right_now_as_api <- date_to_api(right_now)

# and back to current time
api_to_date(right_now_as_api)
}
