% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_url.R
\name{construct_url}
\alias{construct_url}
\title{Construct a URL to access Reddit Content}
\usage{
construct_url(type = "comment", author = "", subreddit = "", size = 500,
  before = as.numeric(Sys.time()), manual = NULL)
}
\arguments{
\item{type}{Type of content you're looking for. Permissible values are "comment" and "submission"}

\item{author}{String representing the handle of a specific Reddit account}

\item{subreddit}{String representing a subreddit to extract data from}

\item{size}{How many records to return, max is 500}

\item{before}{Integer represting time in . Defaults to current time. Functionality for using this parameter more comprehensively in development.}

\item{manual}{String that lets you interface directly with the API. Overrides any other parameters and appends passed argument to URL.}
}
\value{
A string representing a URL providing Reddit content
}
\description{
Construct a URL to access Reddit Content
}
\examples{
#Grab 500 most recent comments on Reddit
construct_url()
}
