% Generated by roxygen2 (4.0.2): do not edit by hand
\name{login_reddit}
\alias{login_reddit}
\title{Login to reddit}
\usage{
login_reddit(user, pass, client, secret)
}
\arguments{
\item{user}{user name}

\item{pass}{password for user}

\item{client}{client ID}

\item{secret}{secret key}
}
\description{
Returns a token needed for API commands that require oauth
}

