% Generated by roxygen2 (4.0.2): do not edit by hand
\name{post_comment}
\alias{post_comment}
\title{Post comment to thread}
\usage{
post_comment(txt, id, token)
}
\arguments{
\item{txt}{Text of comment, formated in markdown}

\item{id}{id of reddit thread to post to}

\item{token}{oauth token generated from login_reddit}
}
\description{
Posts a top level comment to a reddit thread
}

